Move SmallAccountID into common
