+++ shared/wallets/common/index.js
@@ -3,3 +3,4 @@ export {default as AccountEntry} from './account-entry'
 export {default as ParticipantsRow} from './participants-row'
 export {default as WalletPopup} from './wallet-popup'
 export {default as EnterNamePopup} from './enter-name-popup'
+export {default as SmallAccountID} from './small-account-id'
